layout/generic/test/file_scroll_position_iframe.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1305579
-->
<head>
  <meta charset="utf-8">
  <meta name="viewport" content="width=device-width,initial-scale=1">
  <title>Test for Bug 1305579</title>
  <script src="/tests/SimpleTest/paint_listener.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1305579">Mozilla Bug 1305579</a>
<p id="display"></p>
<style>
.off {
  display:none;
}
</style>
<script>
const is = window.opener.is;
const SimpleTest = window.opener.SimpleTest;
const original_finish = window.opener.SimpleTest.finish;

SimpleTest.finish = function finish() {
  self.close();
  original_finish();
}

function runtest() {
  var iframe = document.getElementById("iframe");
  iframe.contentDocument.scrollingElement.scrollTop = 50;
  iframe.classList.toggle("off");
  waitForAllPaintsFlushed(function() {
    iframe.classList.toggle("off");
    is(iframe.contentDocument.scrollingElement.scrollTop, 50, "scroll position restored");
    SimpleTest.finish();
  });
}
</script>
<iframe onload="runtest()" id="iframe" class="" srcdoc="<p>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br>bla<br></p>"></iframe><br>
</body>
</html>
